# perf: optimize pg describe tables query